Add logging with sl4j
In this commit the application has been moved to a single package called bpelconverter. sl4j is added as logger for both spring and the application itself.
