--- src/main/java/models/bpel/Process.java
@@ -1,5 +1,0 @@
-package models.bpel;
-
-public class Process {
-
-}
