--- src/main/java/models/bpel/Scope.java
@@ -1,5 +1,0 @@
-package models.bpel;
-
-public class Scope extends Process {
-    // TODO: Implement this. A scope is like subprocess, need to investigate if it can extend a Process.
-}
