--- src/main/java/models/bpel/activities/basic/Assign.java
@@ -1,16 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-// TODO: Implement
-public class Assign extends Activity {
-
-    public Assign(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null;
-    }
-}
