--- src/main/java/models/bpel/activities/basic/Exit.java
@@ -1,16 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-public class Exit extends Activity {
-
-    public Exit(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null; // TODO
-    }
-
-}
