--- src/main/java/models/bpel/activities/basic/Invoke.java
@@ -1,16 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-// TODO: Implement; Maybe create a super class for all message events?
-public class Invoke extends Activity {
-
-    public Invoke(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null;
-    }
-}
