--- src/main/java/models/bpel/activities/basic/Receive.java
@@ -1,16 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-// TODO: implement
-public class Receive extends Activity {
-
-    public Receive(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null;
-    }
-}
