--- src/main/java/models/bpel/activities/basic/Reply.java
@@ -1,16 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-// TODO: Implement
-public class Reply extends Activity {
-
-    public Reply(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null;
-    }
-}
