--- src/main/java/models/bpel/activities/basic/Rethrow.java
@@ -1,15 +1,0 @@
-package models.bpel.activities.basic;
-
-import models.bpel.activities.Activity;
-
-public class Rethrow extends Activity {
-
-    public Rethrow(String name) {
-        super(name);
-    }
-
-    @Override
-    public Object toBPMN() {
-        return null; // TODO
-    }
-}
