+++ src/main/java/org/bpel2bpmn/models/bpel/Process.java
@@ -1,0 +1,5 @@
+package org.bpel2bpmn.models.bpel;
+
+public class Process {
+
+}
