+++ src/main/java/org/bpel2bpmn/models/bpel/Scope.java
@@ -1,0 +1,5 @@
+package org.bpel2bpmn.models.bpel;
+
+public class Scope extends Process {
+    // TODO: Implement this. A scope is like subprocess, need to investigate if it can extend a Process.
+}
