+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Assign.java
@@ -1,0 +1,16 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+// TODO: Implement
+public class Assign extends Activity {
+
+    public Assign(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null;
+    }
+}
