+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Empty.java
@@ -1,0 +1,16 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+public class Empty extends Activity {
+
+    public Empty(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null; // TODO
+    }
+
+}
