+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Exit.java
@@ -1,0 +1,16 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+public class Exit extends Activity {
+
+    public Exit(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null; // TODO
+    }
+
+}
