+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Receive.java
@@ -1,0 +1,16 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+// TODO: implement
+public class Receive extends Activity {
+
+    public Receive(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null;
+    }
+}
