+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Reply.java
@@ -1,0 +1,16 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+// TODO: Implement
+public class Reply extends Activity {
+
+    public Reply(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null;
+    }
+}
