+++ src/main/java/org/bpel2bpmn/models/bpel/activities/basic/Rethrow.java
@@ -1,0 +1,15 @@
+package org.bpel2bpmn.models.bpel.activities.basic;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+public class Rethrow extends Activity {
+
+    public Rethrow(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null; // TODO
+    }
+}
