+++ src/main/java/org/bpel2bpmn/models/bpel/activities/structured/Flow.java
@@ -1,0 +1,32 @@
+package org.bpel2bpmn.models.bpel.activities.structured;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+import java.util.ArrayList;
+
+public class Flow extends Activity {
+
+    // TODO: Implement Links
+    private ArrayList<Activity> activities;
+
+    public Flow(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null;
+    }
+
+    public void addActivity(Activity activity) {
+        this.activities.add(activity);
+    }
+
+    /*
+     * Getters & Setters
+     */
+
+    public ArrayList<Activity> getActivities() {
+        return activities;
+    }
+}
