+++ src/main/java/org/bpel2bpmn/models/bpel/activities/structured/RepeatUntil.java
@@ -1,0 +1,38 @@
+package org.bpel2bpmn.models.bpel.activities.structured;
+
+import org.bpel2bpmn.models.bpel.activities.Activity;
+
+public class RepeatUntil extends Activity {
+
+    private String condition;
+    private String activity;
+
+    public RepeatUntil(String name) {
+        super(name);
+    }
+
+    @Override
+    public Object toBPMN() {
+        return null;
+    }
+
+    /*
+     * Getters & Setters
+     */
+
+    public String getCondition() {
+        return condition;
+    }
+
+    public void setCondition(String condition) {
+        this.condition = condition;
+    }
+
+    public String getActivity() {
+        return activity;
+    }
+
+    public void setActivity(String activity) {
+        this.activity = activity;
+    }
+}
