+++ src/main/java/org/bpel2bpmn/models/bpmn/Example.java
@@ -1,0 +1,17 @@
+package org.bpel2bpmn.models.bpmn;
+
+import org.camunda.bpm.model.bpmn.Bpmn;
+import org.camunda.bpm.model.bpmn.BpmnModelInstance;
+
+public class Example {
+
+    public static BpmnModelInstance createInstance() {
+        BpmnModelInstance modelInstance = Bpmn.createExecutableProcess("processId")
+                .startEvent()
+                .userTask()
+                .endEvent()
+                .done();
+
+        return modelInstance;
+    }
+}
