+++ src/main/java/org/bpel2bpmn/webservice/Application.java
@@ -1,0 +1,11 @@
+package org.bpel2bpmn.webservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"org.bpel2bpmn.controllers"})
+public class Application {
+    public static void main(String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
