--- src/main/java/utilities/BPELParser.java
@@ -1,16 +1,0 @@
-package utilities;
-
-import models.bpel.Process;
-
-import javax.xml.stream.XMLEventReader;
-import javax.xml.stream.XMLInputFactory;
-
-public class BPELParser {
-    public static Process parse(String bpelXML) {
-//        XMLInputFactory factory = XMLInputFactory.newInstance();
-//        XMLEventReader eventReader =
-//                factory.createXMLEventReader(bpelXML);
-
-        return null; // TODO: Replace with actual return value.
-    }
-}
