+++ src/main/resources/application.yml
@@ -1,11 +1,18 @@
 spring:
   profiles: production
 server:
-  port: 80
+  port: 8080
+logging:
+  level:
+    org.springframework.web: INFO
+    org.bpel2bpmn: INFO
 ---
 
 spring:
   profiles: development
 server:
   port: 8080
-
+logging:
+  level:
+    org.springframework.web: INFO
+    org.bpel2bpmn: DEBUG
